Fix ticket id #55325 & star upscore
"I am not sure for the trigger for this achievement.
I got 75/75 stars in Easy mode, 105/125 in Normal mode and 28/125 in Hard mode.
The same achievement in Sudoku mini triggered when I got a total of 80 over Easy and Normal. " - zer0rc21

The logic was checking for the number of puzzles instead of stars, which actually should've triggered a misfire.

I've updated the logic to be all the stars in every mode by multiplying the puzzle numbers * 5 for the summations, and upscored to reflect a fair amount of work to point yield.